CBF-37 Закрытие диалога
+настройки в Startup
+реализаия MessagePipe при получении сообщения
+Фиксы после тестирования

+++ back/src/Chatbot.Abstractions/Core/Services/ISettingsService.cs
@@ -14,6 +14,6 @@ namespace Chatbot.Abstractions.Core.Services
         Task<bool> Delete(Settings item);
         Task SetDefaultSettings();
         Task<string> GetSalam2();
-        Task<int?> GetClientTimeoutInterval();
+        Task<double?> GetClientTimeoutInterval();
     }
 }
